requirements.txt
scipy==1.5.2
python_dateutil==2.8.1
